Renamed list.c list.h to arraylist.c and arraylist.h
On a new machine, it turns out that check uses several of the
list_* functions in its own code.  Since we are statically
linking in that code that was a problem.

Renaming list_* to arraylist_* more accurately describes an
important detail about the list (that it is array-based).  This
gives us the flexibility in the future to add a linkedlist_*
module without causing too much confusion.